+++ vue-crash-2023/package.json
@@ -10,7 +10,8 @@
   "dependencies": {
     "core-js": "^3.8.3",
     "json-server": "^0.17.3",
-    "vue": "^3.2.13"
+    "vue": "^3.2.13",
+    "vue-router": "^4.0.13"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~5.0.0",
