+++ vue-crash-2023/src/components/Footer.vue
@@ -1,0 +1,17 @@
+<template>
+    <footer>
+        <p>Copyright &copy; 2023</p>
+        <router-link to="/about">About</router-link>
+    </footer>
+</template>
+
+<style scoped>
+a {
+    color: #333;
+}
+
+footer {
+    margin-top: 30px;
+    text-align: center;
+}
+</style>
