+++ vue-crash-2023/src/router/index.js
@@ -1,0 +1,24 @@
+import { createRouter, createWebHistory} from "vue-router";
+
+import Home from '../views/Home'
+import About from '../views/About'
+
+const routes = [
+    {
+        path: '/',
+        name: 'Home',
+        component: Home
+    },
+    {
+        path: '/about',
+        name: 'About',
+        component: About
+    }
+]
+
+const router = createRouter({
+    history: createWebHistory(process.env.BASE_URL),
+    routes
+})
+
+export default router
