+++ vue-crash-2023/src/views/About.vue
@@ -1,0 +1,10 @@
+<template>
+    <h3>About</h3>
+    <router-link to="/">Go Back</router-link>
+</template>
+
+<script>
+export default {
+    
+}
+</script>
